requirements.txt
PyQt5>=5.15
PyMuPDF>=1.23
Pillow>=10.0
numpy>=1.24
opencv-python>=4.8
